+++ internal/config/config.go
@@ -89,8 +89,10 @@ var credentialsChooser = map[string]credentialsDescription{
 		},
 		about: func() string {
 			return fmt.Sprintf(
-				"(deprecated) value of parameter '%s' must be a string with template 'user:password'",
+				"value of parameter %q must be a string with template 'user:password'. Alternatively parameter for splitted parameters %q and %q", //nolint:lll
 				ParamCredentialsStatic,
+				ParamCredentialsStaticLogin,
+				ParamCredentialsStaticPassword,
 			)
 		},
 	},
@@ -100,7 +102,10 @@ var credentialsChooser = map[string]credentialsDescription{
 		},
 		about: func() string {
 			return fmt.Sprintf(
-				"value of parameter '%s' must be a string with static credentials login",
+				"value of parameter %q must be a string with static credentials login. %q uses with %q. Alternatively you can use parameter %q", //nolint:lll
+				ParamCredentialsStaticLogin,
+				ParamCredentialsStaticLogin,
+				ParamCredentialsStaticPassword,
 				ParamCredentialsStatic,
 			)
 		},
@@ -111,7 +116,10 @@ var credentialsChooser = map[string]credentialsDescription{
 		},
 		about: func() string {
 			return fmt.Sprintf(
-				"value of parameter '%s' must be a string with static credentials password",
+				"value of parameter %q must be a string with static credentials password. %q uses with %q. Alternatively you can use parameter %q", //nolint:lll
+				ParamCredentialsStaticPassword,
+				ParamCredentialsStaticPassword,
+				ParamCredentialsStaticLogin,
 				ParamCredentialsStatic,
 			)
 		},